Conflicts:
	liquibase-cdi/pom.xml
	liquibase-core/pom.xml
	liquibase-core/src/main/java/liquibase/logging/LogFactory.java
	liquibase-core/src/main/java/liquibase/parser/core/xml/XMLChangeLogSAXHandler.java
	liquibase-core/src/main/java/liquibase/sqlgenerator/core/CreateTableGenerator.java
	liquibase-core/src/test/java/liquibase/change/core/UpdateDataChangeTest.java
	liquibase-debian/pom.xml
	liquibase-integration-tests/pom.xml
	liquibase-maven-plugin/pom.xml
	liquibase-osgi/pom.xml
	pom.xml

+++ liquibase-core/src/test/java/liquibase/verify/saved_state/minimumRequiredIsValidSql/addColumn/oracle.sql
@@ -4,4 +4,4 @@
 --     type="int"
 -- ], ]
 -- Change Parameter: tableName=person
-ALTER TABLE person ADD id INTEGER;
+ALTER TABLE person ADD id NUMBER(10);
